+++ modules/core-schema/src/main/java/com/enonic/xp/core/impl/schema/relationship/RelationshipTypeLoader.java
@@ -20,7 +20,6 @@ import com.enonic.xp.module.ModuleKey;
 import com.enonic.xp.schema.relationship.RelationshipType;
 import com.enonic.xp.schema.relationship.RelationshipTypeName;
 import com.enonic.xp.schema.relationship.RelationshipTypes;
-import com.enonic.xp.xml.parser.XmlRelationshipTypeParser;
 
 final class RelationshipTypeLoader
 {
