+++ modules/core-schema/src/test/java/com/enonic/xp/core/impl/schema/relationship/XmlRelationshipTypeParserTest.java
@@ -1,10 +1,11 @@
-package com.enonic.xp.xml.parser;
+package com.enonic.xp.core.impl.schema.relationship;
 
 import org.junit.Before;
 import org.junit.Test;
 
 import com.enonic.xp.module.ModuleKey;
 import com.enonic.xp.schema.relationship.RelationshipType;
+import com.enonic.xp.xml.parser.XmlModelParserTest;
 
 import static org.junit.Assert.*;
