+++ modules/core-schema/src/test/resources/com/enonic/xp/core/impl/schema/relationship/XmlRelationshipTypeParserTest.xml
